remove: restore useMypageInfo
